requirements.txt
# ============================================================
# Chatbot Administrativo LSE-FIUBA - Dependencies
# Autor: Juan Ruiz Otondo - CEIA FIUBA
# ============================================================

# Core
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
pydantic>=2.0
pydantic-settings>=2.0

# PDF Processing
PyMuPDF>=1.24.0
pdfplumber>=0.11.0

# Embeddings & Vector Store
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0

# Graph
networkx>=3.3
python-louvain>=0.16
matplotlib>=3.9.0

# LLM
ollama>=0.3.0
openai>=1.30.0

# API
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# UI
streamlit>=1.36.0

# Evaluation
scikit-learn>=1.5.0

# Utilities
python-dotenv>=1.0.0
watchdog>=4.0.0
tqdm>=4.66.0
loguru>=0.7.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# Dev
black>=24.0.0
ruff>=0.5.0
